+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Add `ContainerCreateOptsBuilder::image_os`
 * Add `ContainerCreateOptsBuilder::image_variant`
 * Add `ContainerCreateOptsBuilder::image_arch`
+* Add `ContainerCreateOptsBuilder::manage_password`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync